--- bifrost/proxy/pages/Page.tsx
@@ -1,4 +1,0 @@
-/* This really should not be neccessary. VPS v1 beta turns off client routing if no Page detected - maybe a bug */
-export default function Page() {
-  return null;
-}
