--- bifrost/proxy/pages/wrapped/restorationVisit/route.ts
@@ -1,4 +1,0 @@
-// never matches - see onBeforeRoute for how this route only triggers on restoration visit.
-export default function route() {
-  return false;
-}
