+++ bifrost/renderer/utils/PassthruLayout.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { PropsWithChildren } from "react";
+
+export const PassThruLayout: React.ComponentType<PropsWithChildren> = ({
+  children,
+}) => <>{children}</>;
