+++ bifrost/renderer/wrapped/onRenderHtml.tsx
@@ -1,14 +1,12 @@
 import React from "react";
 import ReactDOMServer from "react-dom/server";
 import { dangerouslySkipEscape, escapeInject } from "vike/server";
-import { PageContextProxyServer } from "../../../types/internal.js";
-import { PageShell } from "../../../lib/PageShell.js";
+import { PageContextProxyServer } from "../../types/internal.js";
+import { PageShell } from "../../lib/PageShell.js";
 import jsdom from "jsdom";
-import { getElementAttributes } from "../../../lib/getElementAttributes.js";
+import { getElementAttributes } from "../../lib/getElementAttributes.js";
 
-export default async function onRenderHtml(
-  pageContext: PageContextProxyServer
-) {
+export async function wrappedOnRenderHtml(pageContext: PageContextProxyServer) {
   if (pageContext.wrappedServerOnly) {
     const { html, layoutProps, layout } = pageContext.wrappedServerOnly;
