+++ tests/vite/pages/_error/+config.ts
@@ -2,6 +2,7 @@ import { BifrostConfig } from "@alignable/bifrost";
 import { MainNavLayout } from "../../layouts/MainNavLayout";
 
 export default {
+  proxyMode: false,
   Layout: MainNavLayout,
   documentProps: {
     title: "Error",
